pyproject.toml
[tool.poetry]
name = "cv-matching"
version = "0.1.0"
description = ""
authors = ["konstantgr <konstantin.grotov@gmail.com>"]
readme = "README.md"
packages = [
    { include = "core" },
    { include = "web" },
]

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
pandas = "^2.1.4"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
torch = "^2.1.2"
transformers = "^4.36.2"
python-docx = "^1.1.0"
scikit-learn = "^1.3.2"
sentence-transformers = "^2.2.2"
tqdm = "^4.66.1"


[tool.poetry.group.web.dependencies]
streamlit = "^1.29.0"
python-docx = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
